Factor out find files
