+++ rust/src/delta_datafusion.rs
@@ -24,13 +24,17 @@ use std::any::Any;
 use std::collections::HashMap;
 use std::convert::TryFrom;
 use std::fmt::Debug;
+use std::pin::Pin;
 use std::sync::Arc;
 
 use arrow::array::ArrayRef;
 use arrow::compute::{cast_with_options, CastOptions};
+use arrow::datatypes::DataType;
 use arrow::datatypes::{DataType as ArrowDataType, Schema as ArrowSchema, SchemaRef, TimeUnit};
 use arrow::error::ArrowError;
 use arrow::record_batch::RecordBatch;
+use arrow_array::StringArray;
+use arrow_schema::Field;
 use async_trait::async_trait;
 use chrono::{DateTime, NaiveDateTime, Utc};
 use datafusion::datasource::datasource::TableProviderFactory;
@@ -43,17 +47,25 @@ use datafusion::optimizer::utils::conjunction;
 use datafusion::physical_expr::PhysicalSortExpr;
 use datafusion::physical_optimizer::pruning::{PruningPredicate, PruningStatistics};
 use datafusion::physical_plan::file_format::FileScanConfig;
+use datafusion::physical_plan::filter::FilterExec;
+use datafusion::physical_plan::limit::LocalLimitExec;
 use datafusion::physical_plan::{
-    ColumnStatistics, ExecutionPlan, Partitioning, SendableRecordBatchStream, Statistics,
+    ColumnStatistics, ExecutionPlan, Partitioning, RecordBatchStream, SendableRecordBatchStream,
+    Statistics,
 };
 use datafusion_common::scalar::ScalarValue;
-use datafusion_common::{Column, DataFusionError, Result as DataFusionResult, ToDFSchema};
+use datafusion_common::tree_node::{TreeNode, TreeNodeVisitor, VisitRecursion};
+use datafusion_common::{
+    Column, DFSchema, DataFusionError, Result as DataFusionResult, ToDFSchema,
+};
+use datafusion_expr::expr::{ScalarFunction, ScalarUDF};
 use datafusion_expr::logical_plan::CreateExternalTable;
-use datafusion_expr::{Expr, Extension, LogicalPlan, TableProviderFilterPushDown};
+use datafusion_expr::{col, Expr, Extension, LogicalPlan, TableProviderFilterPushDown, Volatility};
 use datafusion_physical_expr::execution_props::ExecutionProps;
 use datafusion_physical_expr::{create_physical_expr, PhysicalExpr};
 use datafusion_proto::logical_plan::LogicalExtensionCodec;
 use datafusion_proto::physical_plan::PhysicalExtensionCodec;
+use futures::StreamExt;
 use object_store::ObjectMeta;
 use url::Url;
 
@@ -64,6 +76,8 @@ use crate::table_state::DeltaTableState;
 use crate::{action, open_table, open_table_with_storage_options, SchemaDataType};
 use crate::{DeltaResult, Invariant};
 use crate::{DeltaTable, DeltaTableError};
+
+const PATH_COLUMN: &str = "__delta_rs_path";
 
 impl From<DeltaTableError> for DataFusionError {
     fn from(err: DeltaTableError) -> Self {
@@ -977,6 +991,382 @@ impl TableProviderFactory for DeltaTableFactory {
             open_table_with_storage_options(cmd.to_owned().location, cmd.to_owned().options).await?
         };
         Ok(Arc::new(provider))
+    }
+}
+
+pub(crate) struct FindFilesExprProperties {
+    pub partition_columns: Vec<String>,
+
+    pub partition_only: bool,
+    pub result: DeltaResult<()>,
+}
+
+/// Ensure only expressions that make sense are accepted, check for
+/// non-deterministic functions, and determine if the expression only contains
+/// partition columns
+impl TreeNodeVisitor for FindFilesExprProperties {
+    type N = Expr;
+
+    fn pre_visit(&mut self, expr: &Self::N) -> datafusion_common::Result<VisitRecursion> {
+        // TODO: We can likely relax the volatility to STABLE. Would require further
+        // research to confirm the same value is generated during the scan and
+        // rewrite phases.
+
+        match expr {
+            Expr::Column(c) => {
+                if !self.partition_columns.contains(&c.name) {
+                    self.partition_only = false;
+                }
+            }
+            Expr::ScalarVariable(_, _)
+            | Expr::Literal(_)
+            | Expr::Alias(_, _)
+            | Expr::BinaryExpr(_)
+            | Expr::Like(_)
+            | Expr::ILike(_)
+            | Expr::SimilarTo(_)
+            | Expr::Not(_)
+            | Expr::IsNotNull(_)
+            | Expr::IsNull(_)
+            | Expr::IsTrue(_)
+            | Expr::IsFalse(_)
+            | Expr::IsUnknown(_)
+            | Expr::IsNotTrue(_)
+            | Expr::IsNotFalse(_)
+            | Expr::IsNotUnknown(_)
+            | Expr::Negative(_)
+            | Expr::InList { .. }
+            | Expr::GetIndexedField(_)
+            | Expr::Between(_)
+            | Expr::Case(_)
+            | Expr::Cast(_)
+            | Expr::TryCast(_) => (),
+            Expr::ScalarFunction(ScalarFunction { fun, .. }) => {
+                let v = fun.volatility();
+                if v > Volatility::Immutable {
+                    self.result = Err(DeltaTableError::Generic(format!(
+                        "Find files predicate contains nondeterministic function {}",
+                        fun
+                    )));
+                    return Ok(VisitRecursion::Stop);
+                }
+            }
+            Expr::ScalarUDF(ScalarUDF { fun, .. }) => {
+                let v = fun.signature.volatility;
+                if v > Volatility::Immutable {
+                    self.result = Err(DeltaTableError::Generic(format!(
+                        "Find files predicate contains nondeterministic function {}",
+                        fun.name
+                    )));
+                    return Ok(VisitRecursion::Stop);
+                }
+            }
+            _ => {
+                self.result = Err(DeltaTableError::Generic(format!(
+                    "Find files predicate contains unsupported expression {}",
+                    expr
+                )));
+                return Ok(VisitRecursion::Stop);
+            }
+        }
+
+        Ok(VisitRecursion::Continue)
+    }
+}
+
+pub(crate) struct FindFiles {
+    pub candidates: Vec<Add>,
+    /// Was a physical read to the datastore required to determine the candidates
+    pub parition_scan: bool,
+}
+
+async fn handle_stream(
+    mut stream: Pin<Box<dyn RecordBatchStream + Send>>,
+) -> Result<Option<String>, DeltaTableError> {
+    if let Some(maybe_batch) = stream.next().await {
+        let batch: RecordBatch = maybe_batch?;
+        if batch.num_rows() > 1 {
+            return Err(DeltaTableError::Generic(
+                "Find files returned multiple records for batch".to_owned(),
+            ));
+        }
+        let array = batch
+            .column_by_name(PATH_COLUMN)
+            .unwrap()
+            .as_any()
+            .downcast_ref::<StringArray>()
+            .ok_or(DeltaTableError::Generic(format!(
+                "Unable to downcast column {}",
+                PATH_COLUMN
+            )))?;
+
+        let path = array
+            .into_iter()
+            .next()
+            .unwrap()
+            .ok_or(DeltaTableError::Generic(format!(
+                "{} cannot be null",
+                PATH_COLUMN
+            )))?;
+        return Ok(Some(path.to_string()));
+    }
+
+    Ok(None)
+}
+
+/// Determine which files contain a record that statisfies the predicate
+pub(crate) async fn find_files_scan<'a>(
+    snapshot: &DeltaTableState,
+    store: ObjectStoreRef,
+    schema: Arc<ArrowSchema>,
+    file_schema: Arc<ArrowSchema>,
+    candidates: Vec<&'a Add>,
+    state: &SessionState,
+    expression: &Expr,
+) -> DeltaResult<Vec<&'a Add>> {
+    let mut files = Vec::new();
+    let mut candidate_map: HashMap<String, &'a Add> = HashMap::new();
+
+    let table_partition_cols = snapshot
+        .current_metadata()
+        .ok_or(DeltaTableError::NoMetadata)?
+        .partition_columns
+        .clone();
+
+    let mut file_groups: HashMap<Vec<ScalarValue>, Vec<PartitionedFile>> = HashMap::new();
+    for action in candidates {
+        let mut part = partitioned_file_from_action(action, &schema);
+        part.partition_values
+            .push(ScalarValue::Utf8(Some(action.path.clone())));
+
+        file_groups
+            .entry(part.partition_values.clone())
+            .or_default()
+            .push(part);
+
+        candidate_map.insert(action.path.to_owned(), action);
+    }
+
+    let mut table_partition_cols = table_partition_cols
+        .iter()
+        .map(|c| Ok((c.to_owned(), schema.field_with_name(c)?.data_type().clone())))
+        .collect::<Result<Vec<_>, ArrowError>>()?;
+    // Append a column called __delta_rs_path to track the file path
+    table_partition_cols.push((PATH_COLUMN.to_owned(), DataType::Utf8));
+
+    let input_schema = snapshot.input_schema()?;
+    let input_dfschema: DFSchema = input_schema.clone().as_ref().clone().try_into()?;
+
+    let predicate_expr = create_physical_expr(
+        &Expr::IsTrue(Box::new(expression.clone())),
+        &input_dfschema,
+        &input_schema,
+        state.execution_props(),
+    )?;
+
+    let parquet_scan = ParquetFormat::new()
+        .create_physical_plan(
+            state,
+            FileScanConfig {
+                object_store_url: store.object_store_url(),
+                file_schema,
+                file_groups: file_groups.into_values().collect(),
+                statistics: snapshot.datafusion_table_statistics(),
+                projection: None,
+                limit: None,
+                table_partition_cols,
+                infinite_source: false,
+                output_ordering: None,
+            },
+            None,
+        )
+        .await?;
+
+    let filter: Arc<dyn ExecutionPlan> =
+        Arc::new(FilterExec::try_new(predicate_expr, parquet_scan.clone())?);
+    let limit: Arc<dyn ExecutionPlan> = Arc::new(LocalLimitExec::new(filter, 1));
+
+    let task_ctx = Arc::new(TaskContext::from(state));
+    let partitions = limit.output_partitioning().partition_count();
+    let mut tasks = Vec::with_capacity(partitions);
+
+    for i in 0..partitions {
+        let stream = limit.execute(i, task_ctx.clone())?;
+        tasks.push(handle_stream(stream));
+    }
+
+    for res in futures::future::join_all(tasks).await.into_iter() {
+        let path = res?;
+        if let Some(path) = path {
+            match candidate_map.remove(&path) {
+                Some(action) => files.push(action),
+                None => {
+                    return Err(DeltaTableError::Generic(
+                        "Unable to map __delta_rs_path to action.".to_owned(),
+                    ))
+                }
+            }
+        }
+    }
+
+    Ok(files)
+}
+
+pub(crate) async fn scan_memory_table(
+    snapshot: &DeltaTableState,
+    predicate: &Expr,
+) -> DeltaResult<Vec<Add>> {
+    let actions = snapshot.files().to_owned();
+
+    let batch = snapshot.add_actions_table(true)?;
+    let mut arrays = Vec::new();
+    let mut fields = Vec::new();
+
+    let schema = batch.schema();
+
+    arrays.push(
+        batch
+            .column_by_name("path")
+            .ok_or(DeltaTableError::Generic(
+                "Column with name `path` does not exist".to_owned(),
+            ))?
+            .to_owned(),
+    );
+    fields.push(Field::new(PATH_COLUMN, DataType::Utf8, false));
+
+    for field in schema.fields() {
+        if field.name().starts_with("partition.") {
+            let name = field.name().strip_prefix("partition.").unwrap();
+
+            arrays.push(batch.column_by_name(field.name()).unwrap().to_owned());
+            fields.push(Field::new(
+                name,
+                field.data_type().to_owned(),
+                field.is_nullable(),
+            ));
+        }
+    }
+
+    let schema = Arc::new(ArrowSchema::new(fields));
+    let batch = RecordBatch::try_new(schema, arrays)?;
+    let mem_table = MemTable::try_new(batch.schema(), vec![vec![batch]])?;
+
+    let ctx = SessionContext::new();
+    let mut df = ctx.read_table(Arc::new(mem_table))?;
+    df = df
+        .filter(predicate.to_owned())?
+        .select(vec![col(PATH_COLUMN)])?;
+    let batches = df.collect().await?;
+
+    let mut map = HashMap::new();
+    for action in actions {
+        map.insert(action.path.clone(), action);
+    }
+    let mut files = Vec::new();
+
+    for batch in batches {
+        let array = batch
+            .column_by_name(PATH_COLUMN)
+            .unwrap()
+            .as_any()
+            .downcast_ref::<StringArray>()
+            .ok_or(DeltaTableError::Generic(format!(
+                "Unable to downcast column {}",
+                PATH_COLUMN
+            )))?;
+        for path in array {
+            let path = path.ok_or(DeltaTableError::Generic(format!(
+                "{} cannot be null",
+                PATH_COLUMN
+            )))?;
+            let value = map.remove(path).unwrap();
+            files.push(value);
+        }
+    }
+
+    Ok(files)
+}
+
+pub(crate) async fn find_files<'a>(
+    snapshot: &DeltaTableState,
+    object_store: ObjectStoreRef,
+    schema: Arc<ArrowSchema>,
+    state: &SessionState,
+    predicate: Option<Expr>,
+) -> DeltaResult<FindFiles> {
+    let current_metadata = snapshot
+        .current_metadata()
+        .ok_or(DeltaTableError::NoMetadata)?;
+    let table_partition_cols = current_metadata.partition_columns.clone();
+
+    match &predicate {
+        Some(predicate) => {
+            let file_schema = Arc::new(ArrowSchema::new(
+                schema
+                    .fields()
+                    .iter()
+                    .filter(|f| !table_partition_cols.contains(f.name()))
+                    .cloned()
+                    .collect::<Vec<_>>(),
+            ));
+
+            let input_schema = snapshot.input_schema()?;
+            let input_dfschema: DFSchema = input_schema.clone().as_ref().clone().try_into()?;
+            let expr = create_physical_expr(
+                predicate,
+                &input_dfschema,
+                &input_schema,
+                state.execution_props(),
+            )?;
+
+            // Validate the Predicate and determine if it only contains partition columns
+            let mut expr_properties = FindFilesExprProperties {
+                partition_only: true,
+                partition_columns: current_metadata.partition_columns.clone(),
+                result: Ok(()),
+            };
+
+            TreeNode::visit(predicate, &mut expr_properties)?;
+            expr_properties.result?;
+
+            if expr_properties.partition_only {
+                let candidates = scan_memory_table(snapshot, predicate).await?;
+                Ok(FindFiles {
+                    candidates,
+                    parition_scan: true,
+                })
+            } else {
+                let pruning_predicate = PruningPredicate::try_new(expr, schema.clone())?;
+                let files_to_prune = pruning_predicate.prune(snapshot)?;
+                let files: Vec<&Add> = snapshot
+                    .files()
+                    .iter()
+                    .zip(files_to_prune.into_iter())
+                    .filter_map(|(action, keep)| if keep { Some(action) } else { None })
+                    .collect();
+
+                // Create a new delta scan plan with only files that have a record
+                let candidates = find_files_scan(
+                    snapshot,
+                    object_store.clone(),
+                    schema.clone(),
+                    file_schema.clone(),
+                    files,
+                    state,
+                    predicate,
+                )
+                .await?;
+
+                Ok(FindFiles {
+                    candidates: candidates.into_iter().map(|s| s.to_owned()).collect(),
+                    parition_scan: false,
+                })
+            }
+        }
+        None => Ok(FindFiles {
+            candidates: snapshot.files().to_owned(),
+            parition_scan: true,
+        }),
     }
 }
